Add static-pod-check-workaround utility
The static-pod-check-workaround utility monitors configured static pods.
If any containers are found not to be running for a given pod, it is
considered a health check failure. If any given pod fails the health
check too many consecutive times, kubelet is restarted as a workaround.
The checks are only performed once kubelet has been active for a certain
period.

Signed-off-by: Don Penney <dpenney@redhat.com>