Fix ignition mode and add site-config example
This update fixes the file mode value in the ignition config, which is a
decimal value but was being set as octal. While the value used still
results in the executable bit being set, so the script is able to run,
the other permissions are set incorrectly.

An example site-config has also been added to the README to show where
to place this ignitionConfigOverride yaml, to avoid confusion between
setting it in the cluster vs node level.

Signed-off-by: Don Penney <dpenney@redhat.com>